+++ include/occa/core.hpp
@@ -6,6 +6,7 @@
 #include <occa/core/kernel.hpp>
 #include <occa/core/kernelArg.hpp>
 #include <occa/core/memory.hpp>
+#include <occa/core/memoryPool.hpp>
 #include <occa/core/stream.hpp>
 #include <occa/core/streamTag.hpp>
