+++ include/occa/core/memoryPool.hpp
@@ -9,8 +9,6 @@ namespace occa {
   class modeDevice_t; class device;
   class modeMemoryPool_t;
 
-  namespace experimental {
-
   class memoryPool : public gc::ringEntry_t {
     friend class occa::modeMemoryPool_t;
 
@@ -141,7 +139,7 @@ namespace occa {
      *
      * @endDoc
      */
-    bool operator == (const occa::experimental::memoryPool &other) const;
+    bool operator == (const occa::memoryPool &other) const;
 
     /**
      * @startDoc{operator_equals[1]}
@@ -154,7 +152,7 @@ namespace occa {
      *
      * @endDoc
      */
-    bool operator != (const occa::experimental::memoryPool &other) const;
+    bool operator != (const occa::memoryPool &other) const;
 
     /**
      * @startDoc{resize}
@@ -253,7 +251,6 @@ namespace occa {
     void setAlignment(const udim_t alignment);
   };
 
-  }
 }
 
 #include "memoryPool.tpp"
