+++ include/occa/core/memoryPool.tpp
@@ -1,8 +1,6 @@
 namespace occa {
-namespace experimental {
   template <class T>
   occa::memory memoryPool::reserve(const dim_t entries) {
     return reserve(entries, occa::dtype::get<T>());
   }
-}
 }
