+++ src/occa/internal/core/memoryPool.hpp
@@ -6,7 +6,6 @@
 #include <set>
 
 namespace occa {
-  using experimental::memoryPool;
 
   class modeMemoryPool_t : public modeBuffer_t {
    public:
