+++ tests/src/c/memoryPool.cpp
@@ -2,7 +2,6 @@
 
 #include <occa.hpp>
 #include <occa.h>
-#include <occa/c/experimental.h>
 #include <occa/internal/c/types.hpp>
 #include <occa/internal/utils/testing.hpp>
